experimental: reset KBASEa after chaining a cframe
After linking new cframe to the chain KBASEa still stores the address of
the previous one. If the execution proceeds to <lj_vmeta_call> KBASE
value (i.e. lower 32 bits of the stored address) might be equal to the
current BASE address value so the execution takes the invalid path.

Relates to tarantool/tarantool#4649

Signed-off-by: Igor Munkin <imun@cpan.org>